tests/wpt/web-platform-tests/css/css-grid/alignment/grid-alignment-implies-size-change-030.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Grid Layout Test: Changes on Default-Alignment may affect grid item's width</title>
<link rel="author" title="Javier Fernandez Garcia-Boente" href="mailto:jfernandez@igalia.com">
<link rel="help" href="https://drafts.csswg.org/css-grid-1/#row-align">
<link rel="help" href="https://drafts.csswg.org/css-grid-1/#grid-item-sizing">
<link rel="help" href="https://drafts.csswg.org/css-align-3/#propdef-justify-items">
<link rel="help" href="https://drafts.csswg.org/css-align/#valdef-justify-self-stretch">
<link rel="help" href="https://drafts.csswg.org/css-align/#valdef-justify-self-start">
<meta name="assert" content="Changing the grid's justify-items value from 'start' to 'stretch' causes replaced items to grow.">
<style>
.grid {
  position: relative;
  display: inline-grid;
  grid-template-columns: 200px;
  grid-template-rows: 100px;
  font: 20px/1 Ahem;
  background: grey;
  justify-items: start;
}
#item {
  background: blue;
  justify-self: auto;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<script src="support/style-change.js"></script>
<script>
function runTest() {
  evaluateStyleChange(item, "before", "data-expected-width", 100);
  grid.style.justifyItems = "stretch";
  evaluateStyleChange(item, "after", "data-expected-width", 200);
  done();
}
</script>
<body onload="runTest()">
<div class="grid" id="grid">
  <img data-expected-height="100" id="item" src="support/100x100-green.png"></img>
</div>
</body>
